+++ src/main/java/keeper/project/homepage/user/controller/member/MemberController.java
@@ -12,7 +12,7 @@ import keeper.project.homepage.dto.result.SingleResult;
 import keeper.project.homepage.entity.member.MemberEntity;
 import keeper.project.homepage.service.ResponseService;
 import keeper.project.homepage.user.service.member.MemberDeleteService;
-import keeper.project.homepage.service.posting.PostingService;
+import keeper.project.homepage.user.service.posting.PostingService;
 import keeper.project.homepage.service.util.AuthService;
 import keeper.project.homepage.user.service.member.MemberService;
 import lombok.RequiredArgsConstructor;
